board[x][y], pathClear, presentBoard, getPieceAt
Fixed the bottom rooks' starting positions, fixed the logic in pathClear(), gave presentBoard() numbers and letters as well as a top and bottom to differentiate pieces from the letters on the bottoms, and added a getPieceAt() method. I checked that all of the pieces move correctly (commented out the pawns then just played around in Main) except for the pawns (which I'm checking as I write this and they're actually not working. We should get on that) but other than that everything is fine.